+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - remodeled ngsi-v2 models ([#58,#59,#60](https://github.com/RWTH-EBC/FiLiP/issues/60))
 - improved ContextEntity and Device deleting methods ([#27](https://github.com/RWTH-EBC/FiLiP/issues/28))
 - patch methods for ContextEntity and Device ([#74](https://github.com/RWTH-EBC/FiLiP/issues/74))
+- refactored and improved Examples ([#90](https://github.com/RWTH-EBC/FiLiP/issues/90))
 
 #### v0.1.7
 - introduced automatic testing